# Conflicts:
#	.qmake.conf
#	config.tests/openssl/openssl.cpp
#	mkspecs/features/ctest_testcase_common.prf
#	mkspecs/features/default_pre.prf
#	mkspecs/features/qt_common.prf
#	mkspecs/features/qt_module_headers.prf
#	mkspecs/features/uikit/default_pre.prf
#	mkspecs/win32-clang-msvc/qmake.conf
#	qmake/Makefile.unix
#	qmake/doc/src/qmake-manual.qdoc
#	qmake/library/qmakebuiltins.cpp
#	qmake/qmake.pri
#	src/3rdparty/freetype/qt_attribution.json
#	src/3rdparty/sqlite/qt_attribution.json
#	src/3rdparty/sqlite/sqlite3.c
#	src/3rdparty/sqlite/sqlite3.h
#	src/corelib/configure.json
#	src/corelib/io/qsavefile.cpp
#	src/corelib/io/qtemporaryfile.cpp
#	src/corelib/io/qtemporaryfile_p.h
#	src/corelib/io/qurltlds_p.h
#	src/corelib/json/qjson.cpp
#	src/corelib/json/qjson_p.h
#	src/corelib/json/qjsondocument.cpp
#	src/corelib/tools/UNICODE_LICENSE.txt
#	src/corelib/tools/qlocale.qdoc
#	src/corelib/tools/qlocale_data_p.h
#	src/corelib/tools/qlocale_p.h
#	src/corelib/tools/qt_attribution.json
#	src/corelib/xml/qxmlstream_p.h
#	src/gui/image/qbmphandler.cpp
#	src/gui/kernel/qpalette.cpp
#	src/gui/text/qfont.cpp
#	src/network/access/access.pri
#	src/network/access/qhttpnetworkconnectionchannel_p.h
#	src/plugins/bearer/connman/qconnmanservice_linux.cpp
#	src/plugins/imageformats/gif/qgifhandler.cpp
#	src/plugins/platforms/cocoa/qcocoawindow.mm
#	src/plugins/platforms/cocoa/qnsview.mm
#	src/plugins/platforms/qnx/qqnxrasterwindow.cpp
#	src/plugins/platforms/windows/qwindowsintegration.cpp
#	src/plugins/sqldrivers/sqlite/qsql_sqlite.cpp
#	src/widgets/dialogs/qmessagebox.cpp
#	src/widgets/styles/qwindowsvistastyle.cpp
#	src/widgets/styles/qwindowsxpstyle.cpp
#	src/widgets/widgets/qlabel.cpp
#	src/widgets/widgets/qmenubar.cpp
#	tests/auto/corelib/json/tst_qtjson.cpp
#	tests/auto/corelib/kernel/qobject/tst_qobject.cpp
#	tests/auto/gui/text/qtextlayout/tst_qtextlayout.cpp
#	tests/auto/testlib/selftests/generate_expected_output.py
#	tests/auto/tools/qmakelib/evaltest.cpp
#	tests/auto/widgets/itemviews/qheaderview/tst_qheaderview.cpp
#	tests/auto/widgets/itemviews/qtreeview/tst_qtreeview.cpp

+++ tests/auto/corelib/mimetypes/qmimedatabase/tst_qmimedatabase.cpp
@@ -52,6 +52,7 @@ static const char *const additionalMimeFiles[] = {
     "invalid-magic1.xml",
     "invalid-magic2.xml",
     "invalid-magic3.xml",
+    "magic-and-hierarchy.xml",
     0
 };
 
@@ -72,12 +73,12 @@ static inline QString testSuiteWarning()
     str << "\nCannot find the shared-mime-info test suite\nstarting from: "
         << QDir::toNativeSeparators(QDir::currentPath()) << "\n"
            "cd " << QDir::toNativeSeparators(QStringLiteral("tests/auto/corelib/mimetypes/qmimedatabase")) << "\n"
-           "wget http://cgit.freedesktop.org/xdg/shared-mime-info/snapshot/Release-1-8.zip\n"
-           "unzip Release-1-8.zip\n";
+           "wget http://cgit.freedesktop.org/xdg/shared-mime-info/snapshot/Release-1-10.zip\n"
+           "unzip Release-1-10.zip\n";
 #ifdef Q_OS_WIN
-    str << "mkdir testfiles\nxcopy /s Release-1-8 s-m-i\n";
+    str << "mkdir testfiles\nxcopy /s Release-1-10 s-m-i\n";
 #else
-    str << "ln -s Release-1-8 s-m-i\n";
+    str << "ln -s Release-1-10 s-m-i\n";
 #endif
     return result;
 }
@@ -115,7 +116,6 @@ Q_CONSTRUCTOR_FUNCTION(initializeLang)
 
 static QString seedAndTemplate()
 {
-    qsrand(QDateTime::currentSecsSinceEpoch());
     return QDir::tempPath() + "/tst_qmimedatabase-XXXXXX";
 }
 
@@ -149,7 +149,7 @@ void tst_QMimeDatabase::initTestCase()
     qDebug() << "\nGlobal XDG_DATA_DIRS: " << m_globalXdgDir;
 
     const QString freeDesktopXml = QStringLiteral("freedesktop.org.xml");
-    const QString xmlFileName = QLatin1String(RESOURCE_PREFIX) + freeDesktopXml;
+    const QString xmlFileName = QLatin1String(RESOURCE_PREFIX "packages/") + freeDesktopXml;
     const QString xmlTargetFileName = globalPackageDir + QLatin1Char('/') + freeDesktopXml;
     QVERIFY2(copyResourceFile(xmlFileName, xmlTargetFileName, &errorMessage), qPrintable(errorMessage));
 #endif
@@ -215,6 +215,8 @@ void tst_QMimeDatabase::mimeTypeForName()
 
     QMimeType doesNotExist = db.mimeTypeForName(QString::fromLatin1("foobar/x-doesnot-exist"));
     QVERIFY(!doesNotExist.isValid());
+    QCOMPARE(doesNotExist.comment(), QString());
+    QCOMPARE(doesNotExist.aliases(), QStringList());
 
     // TODO move to findByFile
 #ifdef Q_OS_LINUX
@@ -223,8 +225,10 @@ void tst_QMimeDatabase::mimeTypeForName()
         qWarning() << "ls not found";
     else {
         const QString executableType = QString::fromLatin1("application/x-executable");
+        const QString sharedLibType = QString::fromLatin1("application/x-sharedlib");
         //QTest::newRow("executable") << exePath << executableType;
-        QCOMPARE(db.mimeTypeForFile(exePath).name(), executableType);
+        QVERIFY(db.mimeTypeForFile(exePath).name() == executableType ||
+                db.mimeTypeForFile(exePath).name() == sharedLibType);
     }
 #endif
 
@@ -607,7 +611,7 @@ void tst_QMimeDatabase::allMimeTypes()
     QVERIFY(!lst.isEmpty());
 
     // Hardcoding this is the only way to check both providers find the same number of mimetypes.
-    QCOMPARE(lst.count(), 749);
+    QCOMPARE(lst.count(), 779);
 
     foreach (const QMimeType &mime, lst) {
         const QString name = mime.name();
@@ -636,6 +640,7 @@ void tst_QMimeDatabase::suffixes_data()
     QTest::newRow("mimetype with multiple patterns") << "text/plain" << "*.asc;*.txt;*,v" << "txt";
     QTest::newRow("mimetype with uncommon pattern") << "text/x-readme" << "README*" << QString();
     QTest::newRow("mimetype with no patterns") << "application/x-ole-storage" << QString() << QString();
+    QTest::newRow("default_mimetype") << "application/octet-stream" << QString() << QString();
 }
 
 void tst_QMimeDatabase::suffixes()
@@ -981,6 +986,26 @@ void tst_QMimeDatabase::installNewGlobalMimeType()
         qDebug() << objcsrc.globPatterns();
     }
 
+    const QString fooTestFile = QLatin1String(RESOURCE_PREFIX "magic-and-hierarchy.foo");
+    QCOMPARE(db.mimeTypeForFile(fooTestFile).name(), QString::fromLatin1("application/foo"));
+
+    const QString fooTestFile2 = QLatin1String(RESOURCE_PREFIX "magic-and-hierarchy2.foo");
+    QCOMPARE(db.mimeTypeForFile(fooTestFile2).name(), QString::fromLatin1("application/vnd.qnx.bar-descriptor"));
+
+    // Test if we can use the default comment
+    {
+        struct RestoreLocale
+        {
+            ~RestoreLocale() { QLocale::setDefault(QLocale::c()); }
+        } restoreLocale;
+
+        QLocale::setDefault(QLocale("zh_CN"));
+        QMimeType suseymp = db.mimeTypeForName("text/x-suse-ymp");
+        QVERIFY(suseymp.isValid());
+        QCOMPARE(suseymp.comment(),
+                 QString::fromLatin1("YaST Meta Package"));
+    }
+
     // Now test removing the mimetype definitions again
     for (int i = 0; i < m_additionalMimeFileNames.size(); ++i)
         QFile::remove(destDir + m_additionalMimeFileNames.at(i));
@@ -1043,17 +1068,46 @@ void tst_QMimeDatabase::installNewLocalMimeType()
     QCOMPARE(db.mimeTypeForFile(qmlTestFile).name(),
              QString::fromLatin1("text/x-qml"));
 
-    // Now test removing the local mimetypes again (note, this leaves a mostly-empty mime.cache file)
-    for (int i = 0; i < m_additionalMimeFileNames.size(); ++i)
-        QFile::remove(destDir + m_additionalMimeFileNames.at(i));
+    // Now that we have two directories with mime definitions, check that everything still works
+    inheritance();
+    if (QTest::currentTestFailed())
+        return;
+
+    aliases();
+    if (QTest::currentTestFailed())
+        return;
+
+    icons();
+    if (QTest::currentTestFailed())
+        return;
+
+    comment();
+    if (QTest::currentTestFailed())
+        return;
+
+    mimeTypeForFileWithContent();
+    if (QTest::currentTestFailed())
+        return;
+
+    mimeTypeForName();
+    if (QTest::currentTestFailed())
+        return;
+
+    // Now test removing local mimetypes
+    for (int i = 1 ; i <= 3 ; ++i)
+        QFile::remove(destDir + QStringLiteral("invalid-magic%1.xml").arg(i));
     if (m_isUsingCacheProvider && !waitAndRunUpdateMimeDatabase(m_localMimeDir))
         QSKIP("shared-mime-info not found, skipping mime.cache test");
-    QCOMPARE(db.mimeTypeForFile(QLatin1String("foo.ymu"), QMimeDatabase::MatchExtension).name(),
-             QString::fromLatin1("application/octet-stream"));
-    QVERIFY(!db.mimeTypeForName(QLatin1String("text/x-suse-ymp")).isValid());
+    QVERIFY(!db.mimeTypeForName(QLatin1String("text/invalid-magic1")).isValid()); // deleted
+    QVERIFY(db.mimeTypeForName(QLatin1String("text/x-suse-ymp")).isValid()); // still present
 
-    // And now the user goes wild and uses rm -rf
+    // The user deletes the cache -> the XML provider makes things still work
     QFile::remove(m_localMimeDir + QString::fromLatin1("/mime.cache"));
+    QVERIFY(!db.mimeTypeForName(QLatin1String("text/invalid-magic1")).isValid()); // deleted
+    QVERIFY(db.mimeTypeForName(QLatin1String("text/x-suse-ymp")).isValid()); // still present
+
+    // Finally, the user deletes the whole local dir
+    QVERIFY2(QDir(m_localMimeDir).removeRecursively(), qPrintable(m_localMimeDir + ": " + qt_error_string()));
     QCOMPARE(db.mimeTypeForFile(QLatin1String("foo.ymu"), QMimeDatabase::MatchExtension).name(),
              QString::fromLatin1("application/octet-stream"));
     QVERIFY(!db.mimeTypeForName(QLatin1String("text/x-suse-ymp")).isValid());
